🥅 Increase available patterns capacity from 32 to 64 for WFC and return error when the patterns are more than the capacity

+++ Assets/Scripts/Utils/Wave Function Collapse/Pattern/Creator/ImagePatternSetCreator.cs
@@ -75,7 +75,7 @@ public struct ImagePatternSetCreator
 
                 patternIdByColor.Add(new PatternIdAndColor {
                     Color = pixelData[i],
-                    PatternId = new PatternId((uint) 1 << index)
+                    PatternId = new PatternId((ulong) 1 << index)
                 });
             }
             // I temporarily store the index instead of the actual id (which is the log2(id)) to make list lookup faster
